requirements.txt
aiohttp
websockets
PyJwt